Redesigned version of the ad invoice page
Ad invoice page redesigned as per matts requests.

Linked into the new version of the database. Still requires legit email
account to be added to auto mailers and requires the automated message
text to be updated (waiting till we decide on how we're going to send
out the actual invoice file)