Updated admin page to include testing for email validity
Added ability to test for valid email. uses regex. As we have no
registration system as we thought we originally might, I added what
would have been a part of the registration to the admin page so the
admin can test by putting an email in a text field and clicking a button